Refactor manual GeoJSON upload
